% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/or_not_to_splain.R
\name{or_not_to_splain}
\alias{or_not_to_splain}
\title{Are you mansplaining?}
\usage{
or_not_to_splain(interested = FALSE)
}
\arguments{
\item{interested}{A logical indicating if she asked you to explain something with \emph{genuine} interest. This is not a hypothetical; `interested` is only to be set to `TRUE` if she explicitly asked you to explain something with clear and genuine interest.}
}
\description{
A go-to function for the modern data sci grrl; something to tuck into your handbag before heading out to get your freak on.
}
\examples{
# in general, the default settings usually apply,
or_not_to_splain()

# but, for the occasion where she has indicated clearly that she really does
want to know
or_not_to_splain(interested = TRUE)

}
